Schema v5: Remove blocks.difficulty as it can be calculated from the more compact blocks.bits

+++ mapping/src/mapper/mapper.rs
@@ -37,7 +37,7 @@ impl KaspaDbMapper {
         block.verbose_data.as_ref().expect("Block verbose_data is missing").transaction_ids.len()
     }
 
-    pub fn map_transaction(&self, transaction: &RpcTransaction, subnetwork_key: i16) -> SqlTransaction {
+    pub fn map_transaction(&self, transaction: &RpcTransaction, subnetwork_key: i32) -> SqlTransaction {
         transactions::map_transaction(self.map_payload, subnetwork_key, transaction)
     }
